feat: implement graphics device (closes #253)
