+++ wolf-demo/Cargo.toml
@@ -9,20 +9,19 @@ readme = "README.md"
 repository = "https://github.com/WolfEngine/WolfEngine/"
 version = "3.0.2"
 
-# [lib] # remove this for native local debugging
-# name = "wolf_demo"
-# path = "src/main.rs"
-# crate-type = ["cdylib"]
+[lib] # remove this for native local debugging
+name = "wolf_demo"
+path = "src/main.rs"
+crate-type = ["cdylib"]
 
 [[bin]]
 name = "wolf_demo"
 path = "src/main.rs"
 
 [dependencies]
-anyhow = "*"
+anyhow = {version = "1.0.56", default-features = false}
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-console_error_panic_hook = {version = "0.1.7", default-features = false}
 serde = { version = "1.0.136", default-features = false, features = ["derive"] }
 serde_closure = { version = "0.3.2", default-features = false }
 wasm-bindgen = { version = "0.2.79", default-features = false }
