+++ wolf/src/render/mod.rs
@@ -1,3 +1,2 @@
 pub mod w_graphics_device;
 pub mod w_scene;
-pub mod w_scenes_manager;
